all small coverarts are now 54px converted at the server side
add a bunch of stuff for dealing with displaying Root Folders in the Home view - probably will end up not using, or making an option. needs functional deep nesting in the Music tab.

api forwards in results from getIndexes

we understand receivedIndexes now, but haven't really figured out what to do with it yet