HomeView only uses FadeScroller on tablets
tweaks to FallbackImage to use the Fade stuff by default, but allow it to be overridden, which it is in the Album/Song Item

Reflow on phones when switching between landscape and portrait

Get things to fit better on phones in the Player view

Display the spinner only if our readyState < 3

use AltBarPosition on ProgressSlider to display the streaming progress

Playlist uses scrollTo instead of scrollIntoView

PlaylistsView has a Scroller

add isWideScreen()

generic response (bring up an error popup) when the server spits some kind of an error message to us

server spinner now turned on whenever a server request is sent, and turned off whenever one is received.  Means it's reasonably reliable now.

Fade-ins take longer

add response to onUnload and onWindowRotated